third_party/blink/web_tests/virtual/pna-permission/external/wpt/fetch/private-network-access/mixed-content-fetch.tentative.https.window_include=from-public-expected.txt

This is a testharness.js-based test.
PASS public to local with local targetAddressSpace option: success.
FAIL public to local with local targetAddressSpace option: PUT success. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL public to local with private targetAddressSpace option: failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL public to local with private targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS public to local with public targetAddressSpace option: failed.
PASS public to local with public targetAddressSpace option: PUT failed.
FAIL public to private with local targetAddressSpace option: failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL public to private with local targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS public to private with private targetAddressSpace option: success.
FAIL public to private with private targetAddressSpace option: PUT success. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS public to private with public targetAddressSpace option: failed.
PASS public to private with public targetAddressSpace option: PUT failed.
PASS public to public with local targetAddressSpace option: failed.
FAIL public to public with local targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS public to public with private targetAddressSpace option: failed.
FAIL public to public with private targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS public to public with public targetAddressSpace option: failed.
PASS public to public with public targetAddressSpace option: PUT failed.
Harness: the test ran to completion.
